feat: 增加崩溃统计.

+++ src/main/java/com/zcunsoft/clklog/api/entity/clickhouse/FlowSummarybydate.java
@@ -7,40 +7,73 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 import java.sql.Timestamp;
 
+/**
+ * 浏览概要情况.
+ */
 @Entity(name = "flow_summary_bydate")
 @Data
 public class FlowSummarybydate {
 
+    /**
+     * 统计日期.
+     */
     @Id
     @Column
     Timestamp statDate;
 
+    /**
+     * 渠道.
+     */
     @Id
     @Column
     String lib;
 
+    /**
+     * 项目编码.
+     */
     @Id
     @Column
     String projectName;
+
+    /**
+     * 浏览量.
+     */
     @Column
-    Integer pv;
+    Long pv;
 
+    /**
+     * 访问次数.
+     */
     @Column
-    Integer visitCount;
+    Long visitCount;
 
+    /**
+     * 访客数.
+     */
     @Column
-    Integer uv;
+    Long uv;
 
+    /**
+     * Ip数.
+     */
     @Column
-    Integer ipCount;
+    Long ipCount;
 
+    /**
+     * 访问时长
+     */
     @Column
-    Integer visitTime;
+    Long visitTime;
 
+    /**
+     * 跳出次数.
+     */
     @Column
-    Integer bounceCount;
+    Long bounceCount;
 
-
+    /**
+     * 更新时间.
+     */
     @Column
     Timestamp updateTime;
 }
